+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260422/CHANGELOG.md
@@ -1,4 +1,5 @@
 # nightly
+- Add audio input
 - Add `split_mode` settings
 # V3.6
 - Add Gemma4 support.